requirements.txt
python-nmap
jinja2
matplotlib
pandas
reportlab